pyproject.toml
[tool.poetry]
name = "tvreaders-bot"
version = "0.1.0"
description = ""
authors = ["astashenkov <ya.astashenkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^20.8"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
